python-worker/requirements.txt
insightface==0.7.3
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.2
requests==2.31.0
websocket-client==1.7.0
python-dotenv==1.0.1
colorama==0.4.6